requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
pyarrow==12.0.1
pydeck==0.8.0
Pympler==1.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
smmap==5.0.0
st-pages==0.4.3
streamlit==1.25.0
streamlit-option-menu==0.3.6
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
